+++ Exception/WannaSpeakException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yproximite\WannaSpeakBundle\Exception;
+
+class WannaSpeakException extends \Exception
+{
+
+}
